[Close #216] Move User::Show action.
`Newpoc::Action::User::Show` replaced by
`UsersController::Action::Show`.

RSpec: 720 examples, 0 failures
RuboCop: 244 files inspected, no offenses detected
