--- lib/newpoc-action-user-show/.rspec
@@ -1,2 +1,0 @@
---format documentation
---color
