--- lib/newpoc-action-user-show/Gemfile
@@ -1,4 +1,0 @@
-source 'https://rubygems.org'
-
-# Specify your gem's dependencies in newpoc-action-user-show.gemspec
-gemspec
