--- lib/newpoc-action-user-show/lib/newpoc/action/user/show/version.rb
@@ -1,11 +1,0 @@
-
-module Newpoc
-  module Action
-    module User
-      # Retrieves a user entity from the repository; part of `new_poc`.
-      class Show
-        VERSION = '0.0.2'
-      end
-    end
-  end
-end
